EOS-23085: Created common sub package to be internally leveraged by py-utils
